Refactor to have pyvast at the root
