+++ python/pyvast/__init__.py
@@ -1,0 +1,3 @@
+from .vast import *
+import pyvast.utils.config
+import pyvast.utils.logging
