--- python/vast/__init__.py
@@ -1,3 +1,0 @@
-from .vast import *
-import vast.utils.config
-import vast.utils.logging
